updated files

+++ src/main/java/com/example/demo/Controllers/AuthController.java
@@ -1,10 +1,10 @@
 package com.example.demo.Controllers;
 
-
-import com.example.demo.Model.Users;
+import org.springframework.http.ResponseEntity;
+import com.example.demo.Model.*;
 import com.example.demo.Service.OtpService;
 import com.example.demo.Service.UserService;
-import com.example.demo.dtoUser.LoginReq;
+import com.example.demo.dtoUser.*;
 import com.example.demo.dtoUser.OtpReq;
 import com.example.demo.dtoUser.RegisterRequest;
 import com.example.demo.repo.OTPRepo;
@@ -13,6 +13,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import java.lang.*;
+import java.util.*;
 
 @RestController
 public class AuthController {
@@ -26,28 +28,40 @@ public class AuthController {
     private OTPRepo otpRepo;
 
     @PostMapping("/register")
-    public String register(@RequestBody RegisterRequest request){
-        otpService.register(request.getEmail());
-        return "OTP Sent";
+    public ResponseEntity<?> register(@RequestBody RegisterRequest request){
+//        System.out.println("hi");
+        try{
+            otpService.register(request.getEmail());
+        } catch (RuntimeException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+        return ResponseEntity.ok("OTP Sent");
     }
 
     @PostMapping("/verifyOTP")
-    public String verify(@RequestBody OtpReq otpReq){
+    public Map<String, Object> verify(@RequestBody OtpReq otpReq){
         return otpService.verifyOtp(otpReq);
     }
 
     @PostMapping("/createAccount")
-    public Users createAccount(@RequestBody RegisterRequest request){
+    public ResponseEntity<?> createAccount(@RequestBody RegisterRequest request){
+        if (request.getEmail() == null) {
+            return ResponseEntity.badRequest().body(
+                    Map.of("success", false, "message", "Email is required")
+            );
+        }
         if(!otpService.isVerified(request.getEmail())){
-            throw new RuntimeException("OTP not verified");
+            return ResponseEntity.badRequest().body(Map.of("success",false,
+                    "data","OTP not verified"));
         }
         Users user=userservice.createUser(request);
         otpRepo.delete(otpService.findOtpDetails(request.getEmail()));
-        return user;
+        return ResponseEntity.ok(Map.of("success",true,
+                "data",user));
     }
 
     @PostMapping("/login")
-    public String login(@RequestBody LoginReq req){
+    public LoginRes login(@RequestBody LoginReq req){
         return userservice.verify(req);
     }
